refactor: move schema and types to separate files
